+++ src/runtime/hexagon_remote/host_shim.cpp
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <android/log.h>
 
-#include "bin/src/halide_hexagon_remote.h"
+#include "halide_hexagon_remote.h"
 
 typedef halide_hexagon_remote_handle_t handle_t;
 typedef halide_hexagon_remote_buffer buffer;
@@ -21,7 +21,7 @@ int halide_hexagon_remote_run(handle_t module_ptr, handle_t function,
         int scalar_size = input_scalarsPtrs[i].dataLen;
         if (scalar_size > sizeof(scalar_t)) {
             __android_log_print(ANDROID_LOG_ERROR, "halide", "Scalar argument %d is larger than %d bytes (%d bytes)",
-                                i, sizeof(scalar_t), scalar_size);
+                                i, (int) sizeof(scalar_t), scalar_size);
             return -1;
         }
         memcpy(&scalars[i], input_scalarsPtrs[i].data, scalar_size);
